Add is_positive, is_negative conditions for Number
